requirements.txt
Django
djangorestframework
gunicorn
psycopg2
psycopg2-binary
pyyaml
uritemplate